+++ src/crud/book.py
@@ -140,75 +140,6 @@ class CrudBook:
                 raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Erro ao adicionar um novo status.")
 
 
-    def get_company(self, id_book: int):
-        readers_reading = self.session.query(models.User.id.label('id_user'),
-                                   models.User.nickname,
-                                   models.User.photo,
-                                   models.Status.status.label('status')
-                            )\
-                            .where(and_(models.UserBook.fk_book == id_book, models.UserBook.fk_status == ReadingTypes.READING))\
-                            .join(models.UserBook, models.UserBook.fk_user == models.User.id)\
-                            .join(models.Status, models.Status.id == models.UserBook.fk_status)\
-                            .limit(16).all()
-        readers_read = self.session.query(models.User.id.label('id_user'),
-                                   models.User.nickname,
-                                   models.User.photo,
-                            )\
-                            .where(and_(models.UserBook.fk_book == id_book, models.UserBook.fk_status == ReadingTypes.READ))\
-                            .join(models.UserBook, models.UserBook.fk_user == models.User.id)\
-                            .limit(16).all()
-        readers_to_read = self.session.query(models.User.id.label('id_user'),
-                                   models.User.nickname,
-                                   models.User.photo,
-                            )\
-                            .where(and_(models.UserBook.fk_book == id_book, models.UserBook.fk_status == ReadingTypes.TO_READ))\
-                            .join(models.UserBook, models.UserBook.fk_user == models.User.id)\
-                            .limit(16).all()
-
-        query_status = self.session.query(models.Status).all()
-
-        def find_status(list, value):
-            for x in list:
-                if x.id == value:
-                    return x.status
-            return 'Erro ao carregar status.'
-
- 
-        return {
-            'readers_reading': {
-                'id': ReadingTypes.READING,
-                'status': find_status(query_status, ReadingTypes.READING),
-                'list': readers_reading
-            },
-            'readers_read': {
-                'id': ReadingTypes.READ,
-                'status': find_status(query_status, ReadingTypes.READ),
-                'list': readers_read
-            },
-            'readers_to_read': {
-                'id': ReadingTypes.TO_READ,
-                'status': find_status(query_status, ReadingTypes.TO_READ),
-                'list': readers_to_read,
-            }
-        }
-
-    def get_company_status(self, id_book: int, id_status: int):
-            query = self.session.query(models.User.id.label('id_user'),
-                                   models.User.nickname,
-                                   models.User.photo,
-                            )\
-                            .where(and_(models.UserBook.fk_book == id_book, models.UserBook.fk_status == id_status))\
-                            .join(models.UserBook, models.UserBook.fk_user == models.User.id)\
-                            .all()
-           
-            return {
-            'readers_reading': {
-                'id': id_status,
-                'status': 'lendoa',
-                'list': query
-            }
-        }
-
     #
     # def avaliar_livro(self, id_user, ava: LivroAvaliar):
     #     try:
