+++ src/schemas/usuario.py
@@ -137,3 +137,23 @@ class UsuarioPerfil(UsuarioSimples):
 
 
 UsuarioPerfil.update_forward_refs()
+
+
+class UsersCompanyStatus(BaseModel):
+    id: int
+    status: str
+    readers: Optional[List[UserSuperBasic]] = []
+
+    class Config:
+        orm_mode = True
+        arbitrary_types_allowed = True
+
+
+class UsersCompany(BaseModel):
+    readers_reading: Optional[UsersCompanyStatus] = None
+    readers_read: Optional[UsersCompanyStatus] = None
+    readers_to_read: Optional[UsersCompanyStatus] = None
+
+    class Config:
+        orm_mode = True
+        arbitrary_types_allowed = True
